syntax version to 16bit

+++ src/ti/build.c
@@ -43,7 +43,7 @@ int ti_build_setup(
         uint32_t from_node_id,
         uint8_t from_node_status,
         uint8_t from_node_zone,
-        uint8_t from_node_syntax_ver,
+        uint16_t from_node_syntax_ver,
         uint16_t from_node_port,
         ti_stream_t * stream)
 {
